:heavy_plus_sign: Added parser file with Parser struct to handle manager input
