+++ src/ai.rs
@@ -1,36 +1,23 @@
 use crate::board::Board;
 use crate::model::{CellContent, GameType, Stone};
 use crate::traits::GomokuAI;
-use std::io::Write;
 use std::path::Path;
-use std::process;
-use std::{fs::File, io};
 
-#[derive(Debug)]
 pub struct Ai {
-    pub my_board: Board,
-    pub begin: bool,
-
-    #[cfg(debug_assertions)]
-    pub input_file: File,
-    #[cfg(debug_assertions)]
-    pub output_file: File,
+    pub board: Board,
 }
 
-enum LogType {
-    Unknown,
-    Error,
-    Message,
-    Debug,
-}
+impl Ai {
+    const ABOUT_MAP: [(&'static str, &'static str); 4] = [
+        ("name", "MyAi"),
+        ("version", "0.1"),
+        ("author", "L'eau ça rouille"),
+        ("country", "France"),
+    ];
 
-impl LogType {
-    fn to_str(&self) -> &str {
-        match self {
-            Self::Unknown => "UNKNOWN",
-            Self::Error => "ERROR",
-            Self::Message => "MESSAGE",
-            Self::Debug => "DEBUG",
+    pub fn new() -> Self {
+        Ai {
+            board: Board::new(),
         }
     }
 }
@@ -39,19 +26,19 @@ impl GomokuAI<u8> for Ai {
     fn start(board_size: u8) -> Option<Self> {
         let ai = Ai::new();
 
-        if ai.my_board.size != board_size {
+        if ai.board.size != board_size {
             return None;
         }
         Some(ai)
     }
 
     fn receive_opponent_turn(&mut self, pos: &(u8, u8)) {
-        self.my_board.board[pos.0 as usize][pos.1 as usize] =
+        self.board.board[pos.0 as usize][pos.1 as usize] =
             CellContent::Opponent;
     }
 
     fn play(&mut self) -> (u8, u8) {
-        self.my_board.send_new_pos()
+        self.board.send_new_pos()
     }
 
     fn about(&self) -> &[(&str, &str)] {
@@ -61,16 +48,16 @@ impl GomokuAI<u8> for Ai {
     fn set_board(&mut self, stones: &[(u8, u8, Stone)]) {
         for &(x, y, stone) in stones {
             // Ensure x and y are within bounds of the board
-            if x >= self.my_board.size || y >= self.my_board.size {
+            if x >= self.board.size || y >= self.board.size {
                 continue;
             }
             match stone {
                 Stone::Ally => {
-                    self.my_board.board[x as usize][y as usize] =
+                    self.board.board[x as usize][y as usize] =
                         CellContent::Ally
                 }
                 Stone::Opponent => {
-                    self.my_board.board[x as usize][y as usize] =
+                    self.board.board[x as usize][y as usize] =
                         CellContent::Opponent
                 }
             }
@@ -85,207 +72,4 @@ impl GomokuAI<u8> for Ai {
     fn set_time_left(&mut self, _time: i32) {}
     fn set_game_type(&mut self, _rule: GameType) {}
     fn set_persistent_folder(&mut self, _path: &Path) {}
-}
-
-impl Ai {
-    const ABOUT_MAP: [(&'static str, &'static str); 4] = [
-        ("name", "MyAi"),
-        ("version", "0.1"),
-        ("author", "L'eau ça rouille"),
-        ("country", "France"),
-    ];
-
-    pub fn new() -> Self {
-        let pid = process::id();
-
-        #[cfg(debug_assertions)]
-        let input_file = File::create(format!("input_{}.txt", pid))
-            .expect("Failed to create input file in debug mode");
-        #[cfg(debug_assertions)]
-        let output_file = File::create(format!("output_{}.txt", pid))
-            .expect("Failed to create output file in debug mode");
-
-        Ai {
-            my_board: Board::new(),
-            begin: true,
-
-            #[cfg(debug_assertions)]
-            input_file,
-            #[cfg(debug_assertions)]
-            output_file,
-        }
-    }
-
-    fn handle_about(&mut self, _cmd: &str) -> bool {
-        let bot_name = "my AI";
-
-        println!("name=\"{}\", version=\"1.0\", author=\"Nymand\", country=\"USA\"", bot_name);
-        false
-    }
-
-    fn handle_start(&mut self, cmd: &str) -> bool {
-        let parse: Vec<&str> = cmd.split_whitespace().collect();
-        if let Some(&number_str) = parse.get(1) {
-            if let Ok(size) = number_str.parse::<usize>() {
-                if size != 20 {
-                    println!("ERROR invalid size.");
-                    return false;
-                }
-                println!("OK");
-                return false;
-            }
-        }
-        false
-    }
-
-    fn handle_end(&mut self, _cmd: &str) -> bool {
-        true
-    }
-
-    fn handle_info(&mut self, _cmd: &str) -> bool {
-        false
-    }
-
-    fn handle_begin(&mut self, _cmd: &str) -> bool {
-        self.begin = true;
-        self.my_board.send_new_pos();
-        false
-    }
-
-    fn handle_print(&mut self, _cmd: &str) -> bool {
-        self.my_board.print_board();
-        false
-    }
-
-    fn handle_turn(&mut self, cmd: &str) -> bool {
-        let parse: Vec<&str> = cmd.split_whitespace().collect();
-        let mut x = 0;
-        let mut y = 0;
-
-        if let Some(&number_str) = parse.get(1) {
-            let parts: Vec<&str> = number_str.split(',').collect();
-            if let Some(x_str) = parts.get(0) {
-                if let Ok(size) = x_str.parse::<usize>() {
-                    x = size;
-                }
-            }
-            if let Some(y_str) = parts.get(1) {
-                if let Ok(size) = y_str.parse::<usize>() {
-                    y = size;
-                }
-            }
-        }
-        self.my_board.board[y as usize][x as usize] =
-            CellContent::Opponent;
-        self.my_board.send_new_pos();
-        false
-    }
-
-    fn handle_board(&mut self, _cmd: &str) -> bool {
-        let mut input = String::new();
-        self.my_board.clear_board();
-
-        loop {
-            input.clear();
-            let n = io::stdin().read_line(&mut input);
-            match n {
-                Ok(_) => {}
-                Err(_e) => {
-                    return false;
-                }
-            }
-            if input == "DONE\n" {
-                break;
-            }
-            let parts: Vec<&str> = input.split(',').collect();
-            if parts.len() != 3 {
-                return false;
-            }
-            let x = match parts[0].parse::<usize>() {
-                Ok(num) => num,
-                Err(_) => {
-                    return false;
-                }
-            };
-            let y = match parts[1].parse::<usize>() {
-                Ok(num) => num,
-                Err(_) => {
-                    return false;
-                }
-            };
-            let player = match parts[2].trim_end().parse::<usize>() {
-                Ok(num) => num,
-                Err(_) => {
-                    return false;
-                }
-            };
-            if player == 1 {
-                self.my_board.board[y as usize][x as usize] =
-                    CellContent::Ally;
-            } else {
-                self.my_board.board[y as usize][x as usize] =
-                    CellContent::Opponent;
-            }
-        }
-        self.my_board.send_new_pos();
-        false
-    }
-
-    fn send_log(&self, log_type: LogType, msg: &str) {
-        let log_str = log_type.to_str();
-        println!("{log_str}: {msg}");
-    }
-
-    pub fn handle_command(&mut self, cmd: &str) -> bool {
-        let uppercase =
-            cmd.split_whitespace().next().unwrap().to_uppercase();
-        let token = uppercase.as_str();
-        match token {
-            "ABOUT" => self.handle_about(&cmd),
-            "START" => self.handle_start(&cmd),
-            "END" => self.handle_end(&cmd),
-            "INFO" => self.handle_info(&cmd),
-            "BEGIN" => self.handle_begin(&cmd),
-            "TURN" => self.handle_turn(&cmd),
-            "BOARD" => self.handle_board(&cmd),
-            "PRINT" => self.handle_print(&cmd),
-            _ => {
-                self.send_log(
-                    LogType::Unknown,
-                    "command not implemented",
-                );
-                false
-            }
-        }
-    }
-
-    pub fn start_loop(&mut self) -> std::io::Result<()> {
-        let mut input = String::new();
-
-        loop {
-            input.clear();
-            let n = io::stdin().read_line(&mut input)?;
-
-            self.write_to_input_file(&input);
-
-            if n == 0 {
-                break;
-            }
-
-            if self.handle_command(&input) {
-                break;
-            }
-        }
-        Ok(())
-    }
-
-    pub fn write_to_input_file(&mut self, mes: &str) {
-        #[cfg(debug_assertions)]
-        let _res = write!(self.input_file, "{}", mes);
-    }
-
-    pub fn write_to_output_file(&mut self, mes: &str) {
-        #[cfg(debug_assertions)]
-        let _res = write!(self.output_file, "{}", mes);
-    }
 }
